+++ internal/topic/topicwriterinternal/writer_reconnector_test.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"sort"
 	"sync"
 	"sync/atomic"
 	"testing"
@@ -77,10 +78,10 @@ func TestWriterImpl_CheckMessages(t *testing.T) {
 		maxSize := 5
 		w.cfg.MaxMessageSize = maxSize
 
-		err := w.Write(ctx, []Message{{Data: bytes.NewReader(make([]byte, maxSize))}})
+		err := w.Write(ctx, []PublicMessage{{Data: bytes.NewReader(make([]byte, maxSize))}})
 		require.NoError(t, err)
 
-		err = w.Write(ctx, []Message{{Data: bytes.NewReader(make([]byte, maxSize+1))}})
+		err = w.Write(ctx, []PublicMessage{{Data: bytes.NewReader(make([]byte, maxSize+1))}})
 		require.Error(t, err)
 	})
 }
@@ -141,7 +142,11 @@ func TestWriterImpl_Write(t *testing.T) {
 
 			writeCompleted := make(empty.Chan)
 			go func() {
-				err := e.writer.Write(e.ctx, []Message{{SeqNo: seqNo, CreatedAt: messageTime, Data: bytes.NewReader(messageData)}})
+				err := e.writer.Write(e.ctx, []PublicMessage{{
+					SeqNo:     seqNo,
+					CreatedAt: messageTime,
+					Data:      bytes.NewReader(messageData),
+				}})
 				require.NoError(t, err)
 				close(writeCompleted)
 			}()
@@ -187,7 +192,7 @@ func TestWriterImpl_WriteCodecs(t *testing.T) {
 		})
 
 		require.NoError(t, err)
-		require.NoError(t, e.writer.Write(e.ctx, []Message{{
+		require.NoError(t, e.writer.Write(e.ctx, []PublicMessage{{
 			Data: bytes.NewReader(messContent),
 		}}))
 
@@ -215,7 +220,7 @@ func TestWriterImpl_WriteCodecs(t *testing.T) {
 		})
 
 		require.NoError(t, err)
-		require.NoError(t, e.writer.Write(e.ctx, []Message{{
+		require.NoError(t, e.writer.Write(e.ctx, []PublicMessage{{
 			Data: bytes.NewReader(messContent),
 		}}))
 
@@ -242,7 +247,7 @@ func TestWriterImpl_WriteCodecs(t *testing.T) {
 		codecs := make(map[rawtopiccommon.Codec]empty.Struct)
 
 		for i := 0; i < codecMeasureIntervalBatches; i++ {
-			require.NoError(t, e.writer.Write(e.ctx, []Message{{
+			require.NoError(t, e.writer.Write(e.ctx, []PublicMessage{{
 				Data: bytes.NewReader(messContentShort),
 			}}))
 			// wait send
@@ -251,7 +256,7 @@ func TestWriterImpl_WriteCodecs(t *testing.T) {
 		}
 
 		for i := 0; i < codecMeasureIntervalBatches; i++ {
-			require.NoError(t, e.writer.Write(e.ctx, []Message{{
+			require.NoError(t, e.writer.Write(e.ctx, []PublicMessage{{
 				Data: bytes.NewReader(messContentLong),
 			}}))
 			// wait send
@@ -547,6 +552,38 @@ func TestCreateRawMessageData(t *testing.T) {
 			req,
 		)
 	})
+	t.Run("WithMessageMetadata", func(t *testing.T) {
+		messages := newTestMessagesWithContent(1)
+		messages[0].Metadata = map[string][]byte{
+			"a": {1, 2, 3},
+			"b": {4, 5},
+		}
+		req, err := createWriteRequest(messages, rawtopiccommon.CodecRaw)
+
+		sort.Slice(req.Messages[0].MetadataItems, func(i, j int) bool {
+			return req.Messages[0].MetadataItems[i].Key < req.Messages[0].MetadataItems[j].Key
+		})
+
+		require.NoError(t, err)
+		require.Equal(t, rawtopicwriter.WriteRequest{
+			Messages: []rawtopicwriter.MessageData{
+				{
+					SeqNo: 1,
+					MetadataItems: []rawtopiccommon.MetadataItem{
+						{
+							Key:   "a",
+							Value: []byte{1, 2, 3},
+						},
+						{
+							Key:   "b",
+							Value: []byte{4, 5},
+						},
+					},
+				},
+			},
+			Codec: rawtopiccommon.CodecRaw,
+		}, req)
+	})
 	t.Run("WithSeqno", func(t *testing.T) {
 		req, err := createWriteRequest(newTestMessagesWithContent(1, 2, 3), rawtopiccommon.CodecRaw)
 		require.NoError(t, err)
@@ -711,12 +748,12 @@ func TestCalculateAllowedCodecs(t *testing.T) {
 }
 
 func newTestMessageWithDataContent(num int) messageWithDataContent {
-	res := newMessageDataWithContent(Message{SeqNo: int64(num)}, testCommonEncoders)
+	res := newMessageDataWithContent(PublicMessage{SeqNo: int64(num)}, testCommonEncoders)
 	return res
 }
 
-func newTestMessages(numbers ...int) []Message {
-	messages := make([]Message, len(numbers))
+func newTestMessages(numbers ...int) []PublicMessage {
+	messages := make([]PublicMessage, len(numbers))
 	for i, num := range numbers {
 		messages[i].SeqNo = int64(num)
 	}